Compilation and test fixes for Ubuntu 15.04

+++ file_stream.cc
@@ -83,7 +83,7 @@ Rwsh_ostream& File_ostream::operator<<(double r) {
 
 Rwsh_ostream& File_ostream::operator<<(struct timeval r) {
   if (!dest) open();
-  if (fprintf(dest, "%ld.%06d", r.tv_sec, r.tv_usec) < 0)
+  if (fprintf(dest, "%ld.%06ld", (long) r.tv_sec, (long) r.tv_usec) < 0)
     std::cerr <<"failed fprintf with errno " <<errno <<std::endl;
   return *this;}
